Make the destinction between, key, entry, and match objects clearer.
Update documentation to include a key_t type with key_hash() and
key_cmp() methods and show how entry_t and match_t should "inherit"
from it.

Rename all args '*e' to '*k' where a key type is sufficient.

Flip the argument order to cmp() functions so the first argument is
the key or mutable match object to search for, and the second is the
element being compared. This makes match_cmp() functions more "OO".

Change the argument order to hashtable_index() so the cmp fuction
comes before the *k key/match object. It feels neater.

Rename "same_cmp()" to "pointer_cmp()" to be clearer what it does.

Rename the hashtable_iter_t 'hasht' field to 'htable'. It seems
better.

Update tests to include a key_t type with entry_t and match_t
"inheriting" from it.