In hashtable.[ch] add a ktable containing hash keys.
Having a separate key table means we can quickly probe for matching
keys with good cache locality without referencing the elements. We use
key=0 as a marker for empty buckets so we don't need to reference the
element table to check if the element pointer is NULL.

Change the max load factor from 0.5 to 0.8. With good cache locality
probing higher load factors still produce great performance.

This literally doubled the performance of the cache.