Rename rs_get_filesize() to rs_file_size() and export it.
This is the last fileutil.c function that's not publicly exported, and it's
useful as a platform independent size function that works with the other file
utility functions.

Change it to be a simple function that returns the size. Change the one
place that uses it. Remove the now redundant fileutil.h. Update the docstrings
for all the rs_file_*() functions to be more helpful.