Create a script to generate a class and import overview
It's a simple shell script and it could be done better (and I know there
are some tools to do such analysis), but I wanted to get a first impression
of where we stand. The summary is that the class structure is perhaps not
too bad but the imports are all over the place.

To generate the graphs yourself, you'll need plantuml (see the script comments).