Improve hardlink tests
Include checks of the mirror metadata files to assure that they contain the
correct number of links for hardlinked files and that they have a sha1 hash for
the first (and only the first) file in a series of hardlinked files.

Also add more descriptive comments of what is being tested and change names
used in the code to be more descriptive.