Improve hardlink handling to have correct count, restore and hash (#240)
This patch was originally submitted about 10 years ago.  Details can be
found here: https://savannah.nongnu.org/bugs/?26848
The patch has been by in use for some time, e.g., Debian:
https://salsa.debian.org/python-team/applications/rdiff-backup/blob/7cd963df/debian/patches/03_fix_hardlinks.diff
FIX: added and moved hardlinks were not correctly counted and restored, Closes #239 
FIX: rdiff-backup complained about missing SHA checksums of hardlinks, Closes #78

The hardlink tests have also been revised so as to agree with the changes in
the patch. Include checks of the mirror metadata files to assure that they contain the
correct number of links for hardlinked files and that they have a sha1 hash for
the first (and only the first) file in a series of hardlinked files.