Workaround for when the backup directory is under the source directory
FIX: add workaround to avoid error when backup directory is under the source directory (see issue #296), there is a warning but the backup can succeed.

The underlying issue is that select_mirror can't contain a selection object because
it isn't pickable, but the function Select can't be called on a selection iterable.