Introduce Python 3.9 and remove Python 3.5
CHG: rdiff-backup now supports the newly released Python 3.9 and stops supporting the obsolete Python 3.5.
It's not completely true now but it'll be true once the changelog is written on this basis.
The order of versions in Travis-CI has also been changed to have the supported Windows version be first;
I think this will allow to simplify the setup in an upcoming step.