Add version to schema and use format placeholders
NEW: the current rdiff-backup version can be used in the remote schema with 'x.y.z' being split as placeholders '{vx}', '{vy}', '{vz}' so that one can install (via pip) and use a specific major/minor version of rdiff-backup (see the migration docs for details).
CHG: the host placeholder in the remote schema is now '{h}', '%s' is deprecated.