Implement forced regress
FIX: regress option in API 201 was actually still using the old API function

NEW: (API 201 only) regress action can be forced on a sane repository, i.e. the last backup can be removed, closes #10

CHG: start to introduce more specific return codes (1 for error, 2 for warning, 4 for single file error, etc), which can be combined (e.g. 3 for error _and_ warning), this will take time to introduce everywhere

It's a bit mixing different topics but this became necessary to
differentiate the rejection of forced regress, which is a warning,
from an actual failure to regress