Improve error message when temporary directory is full
FIX: add message about temporary directory to no space left on device, this impacts only cases where tempdir isn't explicitly set, closes #624

When the tempdir is explicitly set, there are much more places where it is used, and the location information isn't carried by the exception so that it would be very difficult to find out this information.
Added accordingly a note to the manpage.