Fix CloseConnections doesn't wait for child to die (#837)
FIX: CloseConnections doesn't wait for child prcesses to die, losing output, closes #819

Keep track of the process object within the LowLevelPipeConnection so we
can wait() for it properly when closing the connection, thus avoiding
possibly lost output from commands run in the remote-schema after the
rdiff-backup server process.

Remove the old _process global that tracked the process purely for
avoiding the "ResourceWarning: subprocess N is still running". As long
as we are pretty sure the process is dead/killed before the connection
destructor, the warning should not recur.  The extra terminate/kill and
sleeps should not run in the normal case, but if delays occur, should
help to ensure the process is marked dead before destruction.

- refactor slightly to add process concept to PipeConnection not LowLevel
- add log messages when process is terminated resp. killed
- add test with long sleeping process

Co-authored-by: Trevor Cordes <git@tecnopolis.ca>