Call msvcrt.setmode only when main is called (#885) (#886)
FIX: Call msvcrt.setmode under Windows only when main is called, allowing for stdout and stderr to be captured by testing framework like pytest, closes #885 

When stdout and stderr are capture by testing framework like pytest,
fileno is not implemented and raise an error. Which prevent the module
to be loaded. To avoid this problem, make the call in main function.