Improve regression speed (#951)
NEW: rdiff-backup uses metadata like the checksum to validate the need for a regression, greatly improving speed of regression.

To enhance the speed of regression, consider verifying the original file's metadata instead of restoring the data. This approach accelerates the regression process by validating the file signature without the need for a full restoration. Given that reading is typically faster than writing, this significantly improves the efficiency of the regression process.