Add timeout when reading file not on disk
When using clound file storage, some time the file is not present on
disk and the software is suposed to fetch the data from the cloud in
real-time. When this process is failing, this cause rdiff-backup to be
stuck trying to read a file indefinitely. To avoid this, a timeout is
added to read the first bytes of the file.